fix: move filter handling to nocodb-sdk
