+++ packages/nocodb-sdk/src/lib/errorUtils.ts
@@ -1,0 +1,7 @@
+export class NcSDKError extends Error {
+  constructor(message: string) {
+    super(message);
+  }
+}
+
+export class BadRequest extends NcSDKError {}
