+++ packages/nocodb-sdk/src/lib/index.ts
@@ -32,3 +32,5 @@ export * from '~/lib/dateTimeHelper';
 export * from '~/lib/form';
 export * from '~/lib/aggregationHelper';
 export * from '~/lib/connectionConfigUtils';
+export * from '~/lib/filterHelpers';
+export * from '~/lib/errorUtils';
